+++ packages/db/src/schema/index.ts
@@ -10,3 +10,4 @@ export * from "./statutory_parameters";
 export * from "./payroll_runs";
 export * from "./payroll_items";
 export * from "./document_vault";
+export * from "./hr_policies";